19.12.2020 progress commit

+++ categoryWindow.py
@@ -5,7 +5,7 @@ from ui.dialogchangecategory import Ui_Dialog as Ui_DialogChange
 from dbConnecter import DbConnector
 
 
-class CategoryWindow(QtWidgets.QDialog):
+class CategoryWindow(DbConnector, QtWidgets.QDialog):
     def __init__(self, parent=None):
         super().__init__(parent)
 
@@ -13,7 +13,11 @@ class CategoryWindow(QtWidgets.QDialog):
         self.ui.setupUi(self)
 
         self.setWindowTitle("Kategorien bearbeiten")
-        self.fillTableWidgetCat()
+
+        if self.serverIsOn() == True:
+            self.fillTableWidgetCat()
+        else:
+            self.msgDbCritical()
 
         self.ui.btnSave.clicked.connect(self.writeDB)
         self.ui.btnUpdate.clicked.connect(self.fillTableWidgetCat)
@@ -27,28 +31,36 @@ class CategoryWindow(QtWidgets.QDialog):
     def writeDB(self):
         category = self.ui.lineEditCategory.text()
         typ = self.ui.comboBoxInOrOutcome.currentText()
-        conn = DbConnector().connect()
-        cur = conn.cursor()
-        cur.execute("""INSERT INTO category (category_id, category, in_or_outcome) VALUES (default, '{0}', '{1}')""".format(category, typ))
-        conn.commit()
+        if self.serverIsOn() == True:
+            conn = DbConnector().connect()
+            cur = conn.cursor()
+            cur.execute("""INSERT INTO category (category_id, category, in_or_outcome) VALUES (default, '{0}', '{1}')""".format(category, typ))
+            conn.commit()
+        else:
+            self.msgDbCritical()
 
         self.ui.lineEditCategory.clear()
         self.fillTableWidgetCat()
 
     def fillTableWidgetCat(self):
-        conn = DbConnector().connect()
-        cur = conn.cursor()
-        cur.execute("SELECT * FROM category ORDER BY category_id ASC")
-        result = cur.fetchall()
-        self.ui.tableWidget.setRowCount(0)
-        for rowNumber, rowData in enumerate(result):
-            self.ui.tableWidget.insertRow(rowNumber)
-            self.ui.tableWidget.setItem(rowNumber, 0, QtWidgets.QTableWidgetItem(str(rowData[0])))
-            self.ui.tableWidget.setItem(rowNumber, 1, QtWidgets.QTableWidgetItem(str(rowData[1])))
-            self.ui.tableWidget.setItem(rowNumber, 2, QtWidgets.QTableWidgetItem(str(rowData[2])))
+        if self.serverIsOn() == True:
+            conn = DbConnector().connect()
+            cur = conn.cursor()
+            cur.execute("SELECT * FROM category ORDER BY category_id ASC")
+            result = cur.fetchall()
+            self.ui.tableWidget.setRowCount(0)
+            for rowNumber, rowData in enumerate(result):
+                self.ui.tableWidget.insertRow(rowNumber)
+                self.ui.tableWidget.setItem(rowNumber, 0, QtWidgets.QTableWidgetItem(str(rowData[0])))
+                self.ui.tableWidget.setItem(rowNumber, 1, QtWidgets.QTableWidgetItem(str(rowData[1])))
+                self.ui.tableWidget.setItem(rowNumber, 2, QtWidgets.QTableWidgetItem(str(rowData[2])))
+        else:
+            self.msgDbCritical()
 
 
-class ChangeCategoryWindow(QtWidgets.QDialog):
+
+
+class ChangeCategoryWindow(DbConnector, QtWidgets.QDialog):
     def __init__(self, id, parent=None):
         super().__init__(parent)
         self.id = id
@@ -64,25 +76,30 @@ class ChangeCategoryWindow(QtWidgets.QDialog):
 
     def windowClose(self):
         self.close()
-        CategoryWindow().fillTableWidgetCat()
 
     def deleteCategory(self):
         id = self.uic.lineEditId.text()
-        conn = DbConnector().connect()
-        cur = conn.cursor()
-        cur.execute("""DELETE FROM category where category_id = '{0}'""".format(id))
-        conn.commit()
-        self.windowClose()
+        if self.serverIsOn() == True:
+            conn = DbConnector().connect()
+            cur = conn.cursor()
+            cur.execute("""DELETE FROM category where category_id = '{0}'""".format(id))
+            conn.commit()
+            self.windowClose()
+        else:
+            self.msgDbCritical()
 
     def updateCategory(self):
         category = self.uic.lineEditCategory.text()
         typ = self.uic.comboBoxTyp.currentText()
         id = self.uic.lineEditId.text()
-        conn = DbConnector().connect()
-        cur = conn.cursor()
-        cur.execute("""UPDATE category SET category = '{0}', in_or_outcome = '{1}' WHERE category_id = '{2}'""".format(category, typ, id))
-        conn.commit()
-        self.windowClose()
+        if self.serverIsOn() == True:
+            conn = DbConnector().connect()
+            cur = conn.cursor()
+            cur.execute("""UPDATE category SET category = '{0}', in_or_outcome = '{1}' WHERE category_id = '{2}'""".format(category, typ, id))
+            conn.commit()
+            self.windowClose()
+        else:
+            self.msgDbCritical()
 
     def changeCategory(self, id):
         conn = DbConnector().connect()
@@ -97,6 +114,8 @@ class ChangeCategoryWindow(QtWidgets.QDialog):
             self.uic.comboBoxTyp.setCurrentIndex(0)
         else:
             self.uic.comboBoxTyp.setCurrentIndex(1)
+
+
